# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Stock Portfolio API",
        "description": "A simple api for managing a user's stock portfolio."
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/sell": {
            "put": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Specifies a stock ticker and a number of shares to sell",
                "operationId": "portfolio_sell",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "ticker",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "format": "int32"
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        },
...
    },
...
}